changed push_* result

+++ inc/rb.h
@@ -6,7 +6,7 @@
 /*   By: npineau <npineau@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/05/23 09:39:35 by npineau           #+#    #+#             */
-/*   Updated: 2017/11/28 14:39:09 by npineau          ###   ########.fr       */
+/*   Updated: 2017/11/28 15:27:28 by npineau          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,30 +39,30 @@ int				rb_force_enqueue(t_rb *rb, void (*del)(void*), void const *src);
 void			rb_free(t_rb *rb);
 void			*rb_front(t_rb rb);
 int				rb_full(t_rb rb);
-int				rb_grow_push_back(t_rb *rb, void const *elem);
-int				rb_grow_push_front(t_rb *rb, void const *elem);
+size_t			rb_grow_push_back(t_rb *rb, void const *elem);
+size_t			rb_grow_push_front(t_rb *rb, void const *elem);
 void			*rb_index(t_rb rb, size_t i);
 void			rb_iter(t_rb rb, void (*f)(void *));
 t_rb			*rb_map(t_rb rb, size_t nsize, void (*f)(void *in, void *out),
 		t_rb *out);
 int				rb_new(size_t capacity, size_t esize, t_rb *slot);
 int				rb_peek(t_rb rb, void **item);
-int				rb_pop_back(t_rb *rb, void *item);
-int				rb_pop_front(t_rb *rb, void *item);
-int				rb_push_back(t_rb *rb, void const *src);
-int				rb_push_back_n(t_rb *rb, void const **src, size_t n);
-int				rb_push_back_n_with(
+size_t			rb_pop_back(t_rb *rb, void *item);
+size_t			rb_pop_front(t_rb *rb, void *item);
+size_t			rb_push_back(t_rb *rb, void const *src);
+size_t			rb_push_back_n(t_rb *rb, void const **src, size_t n);
+size_t			rb_push_back_n_with(
 		void (*cpy)(void const *in, void *out, size_t size),
 		t_rb *rb, void const **src, size_t n);
-int				rb_push_back_with(
+size_t			rb_push_back_with(
 		void (*cpy)(void const *in, void *out, size_t size),
 		t_rb *rb, void const *src);
-int				rb_push_front(t_rb *rb, void const *src);
-int				rb_push_front_n(t_rb *rb, void const **src, size_t n);
-int				rb_push_front_n_with(
+size_t			rb_push_front(t_rb *rb, void const *src);
+size_t			rb_push_front_n(t_rb *rb, void const **src, size_t n);
+size_t			rb_push_front_n_with(
 		void (*cpy)(void const *in, void *out, size_t size),
 		t_rb *rb, void const **src, size_t n);
-int				rb_push_front_with(
+size_t			rb_push_front_with(
 		void (*cpy)(void const *in, void *out, size_t size),
 		t_rb *rb, void const *src);
 t_rb			*rb_resize(t_rb *rb, size_t capacity);
